+++ core/src/test/scala/filodb.core/memstore/TimeSeriesPartitionSpec.scala
@@ -67,7 +67,7 @@ class TimeSeriesPartitionSpec extends MemFactoryCleanupTest with ScalaFutures {
     }
   }
 
-  private val blockStore = new PageAlignedBlockManager(150 * 1024 * 1024,
+  private val blockStore = new PageAlignedBlockManager(200 * 1024 * 1024,
     new MemoryStats(Map("test"-> "test")), reclaimer, 1)
   protected val ingestBlockHolder = new BlockMemFactory(blockStore, None, schema1.data.blockMetaSize,
                                       dummyContext, true)
