+++ query/src/test/scala/filodb/query/exec/BinaryJoinSetOperatorSpec.scala
@@ -21,7 +21,7 @@ import filodb.query._
 
 class BinaryJoinSetOperatorSpec extends FunSpec with Matchers with ScalaFutures {
 
-  import SelectRawPartitionsExecSpec._
+  import MultiSchemaPartitionsExecSpec._
 
   val config = ConfigFactory.load("application_test.conf").getConfig("filodb")
   val queryConfig = new QueryConfig(config.getConfig("query"))
